refactor: upload profile pic
